oamc: Rough addr:place= tag support.
This simply uses the value of the addr:place tag if addr:street is
not present.  Since OsmAnd doesn't associate the address streetnames
with actual streets in the data (unlike Nominatim), the streetname
is simply a string used in address search.  So using the place name
instead still works fine.