Fix zero-length array argument
Reports any call to toArray() on an object of type or subtype
java.util.Collection with a zero-length array argument. When passing in
an array of too small size, the toArray() method has to construct a new
array of the right size using reflection. This has significantly worse
performance than passing in an array of at least the size of the
collection itself.
Powered by InspectionGadgets